pyproject.toml
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "IOU"
version = "0.1.0"
description = ""
authors = ["Josix <wilson8507@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.60.1"
uvicorn = "^0.11.7"
line-bot-sdk = "^1.16.0"
python-dotenv = "^0.14.0"
sqlalchemy = "^1.3.18"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
commitizen = "^2.0.0"
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"
isort = "^5.3.2"
pre-commit = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
